Long shapes.

+++ dl4j-core-benchmark/src/main/java/org/deeplearning4j/memory/BaseMemoryBenchmark.java
@@ -180,7 +180,7 @@ public abstract class BaseMemoryBenchmark {
             }
 
             //Work out output size:
-            int[] outShape;
+            long[] outShape;
             if(mln != null){
                 outShape = mln.output(Nd4j.create(inShape)).shape();
             } else {
